requirements.txt
Django>=4.0,<5.0
djangorestframework
firebird-driver
opencv-python==4.8.1.78
pyzbar==0.1.9
Pillow==10.0.1
numpy==1.24.3